refactor(tui): split pick-sprint, home, and add-ticket views into focused internals
- pick-sprint-view.tsx 589 → 298 LOC; row types, windowing, group builder, and row
  views moved under pick-sprint-internals/. computeWindow stays re-exported at the
  canonical path so the unit-test import keeps resolving.
- home-view.tsx 555 → 254 LOC; StateCard regimes and the menu-item builder split
  out under home-internals/.
- add-ticket-view.tsx 457 → 96 LOC; Step machine, HeaderCard, StepView + runFetch,
  and the ReviewScrollableDescription viewport moved under add-ticket-internals/.

Behaviour preserved (targeted view tests + window unit pass); every file < 350 LOC.
